+++ tests/NetConduit.UnitTests/HandleInitFrameStateGuardTests.cs
@@ -141,9 +141,9 @@ public sealed class HandleInitFrameStateGuardTests
 
         public void NotifyReady(WriteChannel channel) { }
         public void NotifyChannelOpened(string channelId) { }
-        public void NotifyChannelCompleted(ushort channelIndex, string channelId) { }
+        public void NotifyChannelCompleted(uint channelIndex, string channelId) { }
         public void NotifyPendingAcceptCancelled(string channelId) { }
-        public bool SendAck(ushort channelIndex, ulong consumedPosition) => true;
+        public bool SendAck(uint channelIndex, ulong consumedPosition) => true;
         public void NotifyEventHandlerException(Exception exception) { }
         public int PeerMaxRecvPayload => FrameConstants.MaxSlabSize;
